Fuse CSA packed THD layout lowering
Lower fused packed-THD attention indices directly into contiguous original and compressed KV buffers, sharing the CuTeDSL index builder with the CP path. Move the implementation to thd_layout_kernels, retain a legacy CP import wrapper for Lite, and cover sequence-major lowering and real-kernel parity.
