+++ megatron/core/transformer/experimental_attention_variant/csa_utils/fused_sparse_attention.py
@@ -45,6 +45,7 @@ from megatron.core.quantization.indexer_quantization import (
 from megatron.core.tensor_parallel.mappings import async_reduce_scatter_along_first_dim
 from megatron.core.utils import nvtx_range_pop, nvtx_range_push
 
+from . import thd_layout_kernels
 from .csa_teacher_lse import can_use_fused_csa_teacher_lse, fused_csa_teacher_lse
 
 # ---------------------------------------------------------------------------
@@ -2213,6 +2214,10 @@ class FusedCSAIndexerSparseAttnFunc(torch.autograd.Function):
       (``_indexer_topk_core`` THD branch, ``local_to_global_flat`` THD
       branch, ``_compute_dense_*_score`` THD branch,
       ``dense_indexer_backward_wrapper`` with ``cu_seqlens_q/k``).
+      Generic fused THD stores KV as raw ``[all original, all compressed]``
+      sources and lowers the logical indexer output with the shared THD
+      final-index kernel. The legacy per-segment layout remains available to
+      direct callers.
 
     Two indexer-loss variants, selected by the ``sparse_loss`` argument
     (matches ``compute_dsa_indexer_loss`` in the reference ``dsa.py``):
@@ -2242,7 +2247,7 @@ class FusedCSAIndexerSparseAttnFunc(torch.autograd.Function):
         kv_full: Tensor,  # SBHD (skv, b, d) / THD (total_kv_full, d)
         attn_sink: Tensor,  # (np,) f32
         # Window indices (not differentiable)
-        window_idxs: Tensor,  # SBHD (b, sq, win_topk) / THD (total_q, win_topk)
+        window_idxs: Tensor | None,  # SBHD/legacy THD window ids; None for raw THD KV.
         # Indexer inputs (differentiable)
         q_indexer: Tensor,  # SBHD (sq, b, idx_nh, idx_hd) / THD (total_q, idx_nh, idx_hd)
         k_indexer: Tensor,  # SBHD (n_comp, b, idx_hd) / THD (total_comp_idx, idx_hd)
@@ -2268,6 +2273,8 @@ class FusedCSAIndexerSparseAttnFunc(torch.autograd.Function):
         compact_workspace: BSHDCompactIndexerWorkspace | THDCompactIndexerWorkspace | None = None,
         indexer_precision: str = "bf16",
         deterministic: bool = False,
+        thd_window_size: int | None = None,
+        thd_compressed_is_sequence_major: bool = False,
     ) -> Tuple[Tensor, Tensor]:
         """Fused forward: indexer scoring, sparse attention, KL loss, and indexer backward."""
         _ensure_dsa_namespace()
@@ -2347,7 +2354,31 @@ class FusedCSAIndexerSparseAttnFunc(torch.autograd.Function):
         )
 
         # ---- 3. Combine indices (indexer first, then window) + globalize. ----
-        if is_thd:
+        indexer_physical_idxs = None
+        if is_thd and thd_compressed_is_sequence_major:
+            if compressed_kv is None or thd_window_size is None:
+                raise ValueError(
+                    "Raw sequence-major THD lowering requires compressed_kv and thd_window_size."
+                )
+            compressed_base = kv_full.shape[0] - compressed_kv.shape[0]
+            global_idxs, _, indexer_physical_idxs = thd_layout_kernels.build_attention_indices(
+                cu_seqlens_q,
+                0,
+                total_q,
+                0,
+                int(thd_window_size),
+                ratio,
+                indexer_topk,
+                topk_indices_cmp,
+                cu_seqlens_compressed=cu_seqlens_compressed_idx,
+                for_indexer_loss=True,
+                compressed_base=compressed_base,
+                compressed_rows=compressed_kv.shape[0],
+                compressed_is_sequence_major=True,
+            )
+        elif is_thd:
+            if window_idxs is None:
+                raise ValueError("Legacy THD lowering requires caller-supplied window indices.")
             row_batch_ids = batch_of_row(cu_seqlens_q, total_q=total_q)
             offset_per_row = (
                 (cu_seqlens_kv[1:] - cu_seqlens_kv[:-1])[row_batch_ids].unsqueeze(1).to(torch.int32)
@@ -2367,6 +2398,8 @@ class FusedCSAIndexerSparseAttnFunc(torch.autograd.Function):
                 cu_seqlens_kv=cu_seqlens_kv_full,
             )
         else:
+            if window_idxs is None:
+                raise ValueError("SBHD lowering requires caller-supplied window indices.")
             compress_topk_idxs = torch.where(
                 topk_indices_cmp >= 0, topk_indices_cmp + kv_offset, -1
             )
@@ -2467,6 +2500,10 @@ class FusedCSAIndexerSparseAttnFunc(torch.autograd.Function):
                 if compact_predict is not None:
                     compact_predict = compact_predict.clone()
                     compact_predict[padding_row_mask] = 0
+                if indexer_physical_idxs is not None:
+                    indexer_physical_idxs = indexer_physical_idxs.masked_fill(
+                        padding_row_mask.unsqueeze(-1), -1
+                    )
 
             if sparse_loss:
                 if compact_predict is not None:
@@ -2485,12 +2522,15 @@ class FusedCSAIndexerSparseAttnFunc(torch.autograd.Function):
                 # the packed (total_k, D) buffer, so promote per-segment-local
                 # indices to flat-global against cu_seqlens_compressed_idx.
                 if is_thd:
-                    topk_for_target = local_to_global_flat(
-                        topk_indices_cmp,
-                        batch_size=-1,
-                        cu_seqlens_q=cu_seqlens_q,
-                        cu_seqlens_kv=cu_seqlens_compressed_idx,
-                    )
+                    if indexer_physical_idxs is not None:
+                        topk_for_target = indexer_physical_idxs
+                    else:
+                        topk_for_target = local_to_global_flat(
+                            topk_indices_cmp,
+                            batch_size=-1,
+                            cu_seqlens_q=cu_seqlens_q,
+                            cu_seqlens_kv=cu_seqlens_compressed_idx,
+                        )
                 else:
                     topk_for_target = topk_indices_cmp
 
@@ -2596,12 +2636,15 @@ class FusedCSAIndexerSparseAttnFunc(torch.autograd.Function):
                 attn_score_for_bwd = target.clone()
                 index_score_for_bwd = predict.clone()
                 if is_thd:
-                    topk_indices_cmp_global = local_to_global_flat(
-                        topk_indices_cmp,
-                        batch_size=-1,
-                        cu_seqlens_q=cu_seqlens_q,
-                        cu_seqlens_kv=cu_seqlens_compressed_idx,
-                    )
+                    if indexer_physical_idxs is not None:
+                        topk_indices_cmp_global = indexer_physical_idxs
+                    else:
+                        topk_indices_cmp_global = local_to_global_flat(
+                            topk_indices_cmp,
+                            batch_size=-1,
+                            cu_seqlens_q=cu_seqlens_q,
+                            cu_seqlens_kv=cu_seqlens_compressed_idx,
+                        )
                     bwd_q, bwd_w, bwd_k, bwd_attn, bwd_idx, bwd_topk = _thd_to_fake_bshd(
                         q_indexer_flat,
                         w_indexer,
@@ -2783,7 +2826,8 @@ class FusedCSAIndexerSparseAttnFunc(torch.autograd.Function):
         #   cu_seqlens_compressed_idx,
         #   max_seqlen_q, max_seqlen_compressed_idx,
         #   compressed_kv, cu_seqlens_q_unpadded, compact_workspace,
-        #   indexer_precision, deterministic
+        #   indexer_precision, deterministic,
+        #   thd_window_size, thd_compressed_is_sequence_major
         return (
             grad_query,
             grad_kv_full,
@@ -2792,6 +2836,8 @@ class FusedCSAIndexerSparseAttnFunc(torch.autograd.Function):
             grad_q_indexer,
             grad_k_indexer,
             grad_weights,
+            None,
+            None,
             None,
             None,
             None,
@@ -3213,7 +3259,7 @@ def fused_csa_indexer_sparse_attn(
     query: Tensor,
     kv_full: Tensor,
     attn_sink: Tensor,
-    window_idxs: Tensor,
+    window_idxs: Tensor | None,
     q_indexer: Tensor,
     k_indexer: Tensor,
     weights: Tensor,
@@ -3237,6 +3283,8 @@ def fused_csa_indexer_sparse_attn(
     compact_workspace: BSHDCompactIndexerWorkspace | THDCompactIndexerWorkspace | None = None,
     indexer_precision: str = "bf16",
     deterministic: bool = False,
+    thd_window_size: int | None = None,
+    thd_compressed_is_sequence_major: bool = False,
 ) -> Tuple[Tensor, Tensor]:
     """Path B (training): fused indexer (+KL loss) + sparse attention.
 
@@ -3244,12 +3292,11 @@ def fused_csa_indexer_sparse_attn(
 
     * **SBHD** (``cu_seqlens_q is None``, default): inputs carry an
       explicit batch axis; the THD kwargs are ignored.
-    * **THD packed** (``cu_seqlens_q`` supplied): all four
-      ``cu_seqlens_*`` and four ``max_seqlen_*`` must be supplied (see
-      below). Both ``sparse_loss=True`` and ``sparse_loss=False`` are
-      supported — the sparse-loss path globalizes the per-segment-local
-      topk indices via ``local_to_global_flat`` and the cuDNN
-      sparse-indexer-backward kernel addresses K/dK by flat ids.
+    * **THD packed** (``cu_seqlens_q`` supplied): packed-sequence metadata
+      must be supplied as described below. Both ``sparse_loss=True`` and
+      ``sparse_loss=False`` are supported. Raw sequence-major THD lowers top-k
+      directly to physical compressed rows; the legacy per-segment layout
+      retains the eager globalization fallback.
 
     See :class:`FusedCSAIndexerSparseAttnFunc` for the detailed data flow.
 
@@ -3266,11 +3313,11 @@ def fused_csa_indexer_sparse_attn(
 
     THD args (when ``cu_seqlens_q is not None``):
         query:        ``(total_q, np, d)`` bf16 — flat-packed Q.
-        kv_full:      ``(total_kv_full, d)`` bf16 — per-segment-concat'd
-                      ``[kv, compressed_kv]`` (built by
-                      :func:`csa.cat_per_segment`).
-        window_idxs:  ``(total_q, win_topk)`` int32 — local-per-segment
-                      window indices.
+        kv_full:      ``(total_kv_full, d)`` bf16 — either the legacy
+                      per-segment layout or raw ``[all kv, all compressed_kv]``
+                      when ``thd_compressed_is_sequence_major=True``.
+        window_idxs:  ``(total_q, win_topk)`` local window indices for the
+                      legacy layout; pass ``None`` for raw sequence-major THD.
         q_indexer:    ``(total_q, idx_nh, idx_hd)`` bf16.
         k_indexer:    ``(total_comp_idx, idx_hd)`` bf16 — compressed-only K
                       (== Compressor's output, packed flat).
@@ -3278,7 +3325,8 @@ def fused_csa_indexer_sparse_attn(
         kv_offset:    ignored.
         cu_seqlens_q:               ``(B+1,)`` int32 CUDA.
         cu_seqlens_kv:              ``(B+1,)`` int32 — original-KV cu_seqlens.
-        cu_seqlens_kv_full:         ``(B+1,)`` int32 — built by
+        cu_seqlens_kv_full:         ``(B+1,)`` int32 — required only by the
+                                    legacy layout and built by
                                     :func:`csa.build_cu_seqlens_kv_full`.
         cu_seqlens_compressed_idx:  ``(B+1,)`` int32 — Compressor's
                                     second return value.
@@ -3303,9 +3351,8 @@ def fused_csa_indexer_sparse_attn(
             KL is computed over the full causally-valid KV. See
             :class:`FusedCSAIndexerSparseAttnFunc` for the full data flow.
         compressed_kv: THD only (required) — ``(total_compressed_kv, d)``
-            bf16, the pre-packed compressed KV from the Compressor. Used
-            by the loss path; THD ``kv_full`` is per-segment concatenated
-            so it cannot be sliced uniformly the way SBHD ``kv_full`` is.
+            bf16, the sequence-major compressed KV from the Compressor. Used
+            directly by the loss path and as the appended region of raw THD KV.
         calculate_per_token_loss: if True, report raw local KL sum and
             compensate the cuDNN backward wrappers' local averaging.
         cu_seqlens_q_unpadded: THD only (optional) — ``(B+1,)`` int32,
@@ -3323,6 +3370,10 @@ def fused_csa_indexer_sparse_attn(
         deterministic: resolve exact-value compact Top-K ties toward the
             smallest local KV indices. This is normally sourced from
             ``TransformerConfig.deterministic_mode``.
+        thd_window_size: THD raw-layout sliding-window width. Required when
+            ``thd_compressed_is_sequence_major=True``.
+        thd_compressed_is_sequence_major: lower logical compressed ids directly
+            into the appended sequence-major compressed buffer.
     """
     if indexer_precision == "mxfp8" and not sparse_loss and loss_coeff > 0:
         raise ValueError("MXFP8 indexer loss supports only sparse indexer loss")
@@ -3332,7 +3383,6 @@ def fused_csa_indexer_sparse_attn(
             name
             for name, val in (
                 ("cu_seqlens_kv", cu_seqlens_kv),
-                ("cu_seqlens_kv_full", cu_seqlens_kv_full),
                 ("cu_seqlens_compressed_idx", cu_seqlens_compressed_idx),
                 ("max_seqlen_q", max_seqlen_q),
                 ("max_seqlen_compressed_idx", max_seqlen_compressed_idx),
@@ -3340,6 +3390,10 @@ def fused_csa_indexer_sparse_attn(
             )
             if val is None
         ]
+        if not thd_compressed_is_sequence_major and cu_seqlens_kv_full is None:
+            missing.append("cu_seqlens_kv_full")
+        if thd_compressed_is_sequence_major and thd_window_size is None:
+            missing.append("thd_window_size")
         if missing:
             raise ValueError(
                 f"fused_csa_indexer_sparse_attn THD mode requires {missing} " "to all be supplied."
@@ -3371,6 +3425,8 @@ def fused_csa_indexer_sparse_attn(
         compact_workspace,
         indexer_precision,
         deterministic,
+        thd_window_size,
+        thd_compressed_is_sequence_major,
     )
