+++ tests/unit_tests/transformer/experimental_attention_variant/test_attention_variant_csa.py
@@ -2138,7 +2138,7 @@ class TestCompressedSparseAttentionThd:
     @pytest.mark.skipif(not torch.cuda.is_available(), reason="CUDA not available")
     def test_thd_path_a_forward(self):
         """Path A (THD): compress_ratio=128 → indexer=None → attend to
-        ALL compressed positions per segment via ``get_compress_topk_idxs_thd``.
+        all visible compressed positions via final-index lowering.
         """
         compress_ratio = 128
         csa = self._build_csa(compress_ratio)
